Change initialization of SCCI wave function
Previously the coefficients were random numbers between
-0.1 and 0.1. Sometimes important coefficents would end
up with a coefficient of almost 0, so that the corresponding
gradient direction would almost never be sampled, and the
coefficient would never get to update. This leads to a
constant bias in the final result.

Instead, just set the coefficients to a constant value, chosen
from some experiments. This seems to fix this issue and gives a
much better initial WF. In the future this can be improved by
reducing coefficients for higher energy states. Need to find
an appropriate function form for this.