Don't store internally contracted states in class 8 which have 0 connection
Many of the states in excitation class 8 have 0 Hamiltonian matrix
element with the CAS state (even though they're a double excitation
from that state). This can happen if the total spin of i and j is
not the same as a and b, or if we're cutting off elements below
epsilon, or symmetry.

This change means that only internally contracted states which are
connected are stored in the coefficient list.