Improve selection of initial determinants for new SC-NEVPT2 code
Previously, we just found the best determinants connected to the
Hartree--Fock. In some cases, there are no determinants connected
to the HF within a SC space which have a non-zero overlap.

Instead we search for the best initial determinants when sampling
the SC norms. This is cheaper and more effective; any SC spaces
in which we do not find any determinants with non-zero overlaps
will also have a calculated SC norm of 0, so they won't get
sampled anyway, and so this will always work.