E_l^k now calculated with constant residence time, instead of constant iteration count
Also, a new function has been added to calculate the exact E_l^k value
for a given subsapce S_l^k, and then generate a large number of
stochastic samples of E_l^k from this space. This allows the user to
compare the distribution of E_l^k to the exact value, for testing
purposes and investigating possible biases.