Replace boost hash_combine with fixed implementation
This is because different versions of boost are giving different
results from hash_combine, meaning that the tests fail for some
versions. The new code will be consistent. It gives the same result
as before for boost 1.57.0.

(Tests fail becuase this hash value is used to assign determinants
to processors. Changing this mapping changes the random walks).