Implement the SelectedCI wave function for use in FCIQMC as a trial WF
This is useful for using a FCI wave function for testing and sanity
checks.

There are some complications here. First, the appropriate walker
object is now SimpleWalker instead of Walker, so templates have
to be adapated to allow this.

Also, the HamAndOvlp function in SelectedCI.cpp does something
different to HamAndOvlp for other wave functions, so we have to
create a new version of this function to be consistent.