In death routine, remove hash table entry if all walker pops go to 0
A walker is only valid if there is both a hash table entry, and
if the walker population is non-zero (actually more than 1e-12).

Because of the exponential approximation, walkers can end up with
a zero population in the death routine after this is applied. If
the hash table entry is not removed, then this can very rarely
lead to some problems. This commit removes the entry, to prevent
these.

In particular, if a hash table entry is not removed, and then
another determinant is added to slot pointed to by this entry,
and then the original determinant is spawned to again then, because
the hash table entry still exists, it will be added to the other
determinant, which is a bug.